Create Examine_02.py
This script is used to generate some radii-randomly-distributed particles, either from a single origin (corresponding to generate_par_2()) or with the same orientation of motion (corresponding to generate_par_1()). And, further, simulate what they are in WISPR-Inner camera. Maybe there are some duplicate functions in this script that are used in Examine_01 code.